ItemTool.calculateBreakDamage|ItemTool.calculateRightClickDamage return become short, Add Gold Tool, Unfix DiggingHandler

+++ src/main/java/net/glowstone/block/itemtype/ItemHoe.java
@@ -2,7 +2,6 @@ package net.glowstone.block.itemtype;
 
 import net.glowstone.block.GlowBlock;
 import net.glowstone.entity.GlowPlayer;
-
 import org.bukkit.Material;
 import org.bukkit.Sound;
 import org.bukkit.block.BlockFace;
@@ -33,7 +32,7 @@ public class ItemHoe extends ItemTool {
     }
 
     @Override
-    public int calculateBreakDamage(GlowBlock target) {
+    public short calculateBreakDamage(GlowBlock target) {
         return 0;
     }
 }
